django_rest_imageupload_backend/requirements.txt
#EmenuRESTAPI 
#pip freeze > requirements.txt

#REQUIREMENTS:
Django==3.2.13
djangorestframework==3.13.1
asgiref==3.5.2
django-filter==21.1
importlib-metadata==4.11.4
Markdown==3.3.7
Pillow==9.1.1
pytz==2022.1
sqlparse==0.4.2
zipp==3.8.0